requirements.txt
click==8.3.1
colorama==0.4.6
markdown-it-py==4.0.0
mdurl==0.1.2
pdf2image==1.17.0
pillow==12.0.0
Pygments==2.19.2
pypdf==6.3.0
rich==14.2.0
shellingham==1.5.4
typer==0.20.0
typing_extensions==4.15.0
